feat: enhance server configuration and middleware for improved token handling and static file serving
