+++ src/middleware/auth.ts
@@ -14,8 +14,9 @@ export function checkToken(
   next: NextFunction
 ): void {
   const token = req.cookies.token;
-
+  console.log(token);
   if (!token) {
+    console.log(!token);
     return res.redirect('/login.html');
   }
