+++ src/routes/messageRoutes.ts
@@ -1,7 +1,6 @@
 import { Router } from 'express';
 import { MessageController } from '../controllers';
 import { MessageService } from '../services';
-import { checkToken } from '../middleware';
 /**
  * Creates routes for message-related operations.
  *
@@ -12,10 +11,7 @@ export function createMessageRoutes(messageService: MessageService): Router {
   const router = Router();
   const controller = new MessageController(messageService);
 
-  router
-    .route('/')
-    .get(checkToken, controller.getMessages)
-    .post(checkToken, controller.createMessage);
+  router.route('/').get(controller.getMessages).post(controller.createMessage);
 
   return router;
 }
